[FEA] 서비스 필요없는 인터페이스 제거
